Refactoring: Extraction des écrans et amélioration de l'architecture
- Centralisation de la configuration du thème dans AppTheme

- Mise en place du pattern Provider pour la gestion de l'état

- Déplacement des écrans dans des fichiers séparés

- Ajout de la méthode readJsonFile à BackupService
