+++ pubspec.yaml
@@ -25,6 +25,7 @@ dependencies:
   file_picker: ^8.0.0
   share_plus: ^7.2.0
   url_launcher: ^6.2.6
+  provider: ^6.1.1
 
 dev_dependencies:
   flutter_test:
